+++ packages/server/__tests__/db/before-db.js
@@ -1,0 +1,11 @@
+const db = require('../../src/db');
+const fixtures = require('./seeds/fixtures');
+
+before(() => {
+    fixtures.seed(db.knex);
+});
+
+after(() => {
+    console.log('DESTROYING>.............');
+    db.knex.destroy();
+});
