+++ packages/server/package.json
@@ -31,7 +31,7 @@
     "test:agencyimport": "NODE_ENV=test SUPPRESS_EMAIL=true mocha --exit --bail --require __tests__/api/fixtures.js __tests__/db/agencyImporter._test_.js",
     "test:apis": "NODE_ENV=test SUPPRESS_EMAIL=true mocha --exit --bail --require __tests__/api/fixtures.js __tests__/api/*.test.js",
     "test:arpa": "NODE_ENV=test __tests__/run_arpa_reporter_tests.sh",
-    "test:db": "NODE_ENV=test mocha --timeout 10000 __tests__/db/*.test.js",
+    "test:db": "NODE_ENV=test mocha --timeout 10000 __tests__/db/index.js",
     "test:email": "NODE_ENV=test mocha __tests__/email/*.test.js",
     "test:lib": "NODE_ENV=test mocha __tests__/lib/*.test.js",
     "test:userimport": "NODE_ENV=test SUPPRESS_EMAIL=true mocha --exit --bail --require __tests__/api/fixtures.js __tests__/db/userImporter._test_.js",
