+++ packages/server/src/routes/sessions.js
@@ -87,7 +87,8 @@ router.post('/', async (req, res, next) => {
         const passcode = await createAccessToken(email);
         const domain = process.env.API_DOMAIN || process.env.WEBSITE_DOMAIN || req.headers.origin;
         const redirectTo = validatePostLoginRedirectPath(req.body.redirect_to);
-        await sendPassCode(email, passcode, domain, redirectTo);
+        console.log(passcode, domain, redirectTo);
+        // await sendPassCode(email, passcode, domain, redirectTo);
 
         res.json({
             success: true,
